Prepare priors

Takes a list input to allow for modifiable priors. Priors can be specified as an option in run_mod.

Usage

prep_priors(
  max_antigens,
  mu_hyp_param = c(1, 7, 1, -4, -1),
  prec_hyp_param = c(1, 1e-05, 1, 0.001, 1),
  omega_param = c(1, 50, 1, 10, 1),
  wishdf_param = 20,
  prec_logy_hyp_param = c(4, 1)
)

Arguments

max_antigens

An integer specifying how many antigen-isotypes will be modeled

mu_hyp_param

A numeric vector of 5 values representing the mean of the hyperprior distribution of the population mean of the random person-specific seroresponse curve parameters. If specified, must be 5 values long, representing the following parameters:

y0 = baseline antibody concentration (default = 1.0)

y1 = peak antibody concentration (default = 7.0)

t1 = time to peak (default = 1.0)

r = shape parameter (default = -4.0)

alpha = decay rate (default = -1.0)

prec_hyp_param

A numeric vector of 5 values corresponding to diagonal entries representing the precision matrix of the hyperprior distributions of the population means of the seroresponse curve parameters (reused for each antigen-isotype). If specified, must be 5 values long:

defaults: y0 = 1.0, y1 = 0.00001, t1 = 1.0, r = 0.001, alpha = 1.0

omega_param

A numeric vector of 5 values corresponding to the diagonal entries of the (diagonal) scale matrix of the Wishart hyperprior distributions of the person-specific random effects' precision matrices. If specified, must be 5 values long:

defaults: y0 = 1.0, y1 = 50.0, t1 = 1.0, r = 10.0, alpha = 1.0

wishdf_param

An integer vector of 1 value specifying the degrees of freedom for the Wishart hyperprior distributions of the random effects' precision matrices. If specified, must be 1 value long.

default = 20.0

prec_logy_hyp_param

A numeric vector of 2 values corresponding to the hyper-parameters for the precision (inverse variance) of the biomarkers, on log-scale. If specified, must be 2 values long.

defaults = 4.0, 1.0

Value

A "curve_params_priors" object (a subclass of list with the inputs to prep_priors() attached as an attributes entry named "used_priors").

"n_params": Corresponds to the 5 parameters being estimated.

"mu.hyp": A matrix of hyperpriors with dimensions number of antigens x 5 (number of parameters), representing the mean of the hyperprior distribution: y0, y1, t1, r, and alpha).

"prec.hyp": A three-dimensional numeric array with dimensions max_antigens x 5 (# of parameters) x 5 (# of parameters), containing the precision matrices of the hyperprior distributions on the population means of the seroresponse parameters, for each antigen-isotype.

"omega" : A three-dimensional numeric array with 5 matrix,each with dimensions number of antigens x 5 (# of parameters), representing the precision matrix of Wishart hyper-priors

"wishdf": A vector of 2 values specifying Wishart distribution degrees of freedom.

"prec.logy.hyp": A matrix of hyper-parameters for the precision (inverse variance) of the biomarkers, on log-scale, measuring number of antigens x 2.

Examples

prep_priors(max_antigens = 2)
#> $n_params
#> [1] 5
#> 
#> $mu.hyp
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    7    1   -4   -1
#> [2,]    1    7    1   -4   -1
#> 
#> $prec.hyp
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    1    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1]  [,2] [,3] [,4] [,5]
#> [1,]    0 1e-05    0    0    0
#> [2,]    0 1e-05    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    0    0
#> [2,]    0    0    1    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]  [,4] [,5]
#> [1,]    0    0    0 0.001    0
#> [2,]    0    0    0 0.001    0
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    1
#> [2,]    0    0    0    0    1
#> 
#> 
#> $omega
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    1    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0   50    0    0    0
#> [2,]    0   50    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    0    0
#> [2,]    0    0    1    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0   10    0
#> [2,]    0    0    0   10    0
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    1
#> [2,]    0    0    0    0    1
#> 
#> 
#> $wishdf
#> [1] 20 20
#> 
#> $prec.logy.hyp
#>      [,1] [,2]
#> [1,]    4    1
#> [2,]    4    1
#> 
#> attr(,"class")
#> [1] "curve_params_priors" "list"               
#> attr(,"used_priors")
#> attr(,"used_priors")$mu_hyp_param
#> [1]  1  7  1 -4 -1
#> 
#> attr(,"used_priors")$prec_hyp_param
#> [1] 1e+00 1e-05 1e+00 1e-03 1e+00
#> 
#> attr(,"used_priors")$omega_param
#> [1]  1 50  1 10  1
#> 
#> attr(,"used_priors")$wishdf_param
#> [1] 20
#> 
#> attr(,"used_priors")$prec_logy_hyp_param
#> [1] 4 1
#>