Default print for run_mod() output object of class sr_model

A default print method for class sr_model that includes the mean posterior distribution for antibody kinetic curve parameters by Iso_type and Stratification (if specified).

Usage

# S3 method for class 'sr_model'
print(data)

Arguments

data

An sr_model output object from the run_mod() function.

Value

A dplyr::grouped_df that contains the mean posterior distribution for antibody kinetic curve parameters by Iso_type and Stratification (if specified).

Examples

print(nepal_sees_jags_output)
#> # A tibble: 4 × 7
#> # Groups:   Stratification, Iso_type [4]
#>   Stratification Iso_type   alpha shape    t1    y0    y1
#>   <chr>          <chr>      <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 paratyphi      HlyE_IgA 0.00229  1.66  3.85  2.55 1024.
#> 2 typhi          HlyE_IgA 0.00291  1.61  7.31  2.97 1032.
#> 3 paratyphi      HlyE_IgG 0.00257  1.36  4.78  1.72  833.
#> 4 typhi          HlyE_IgG 0.00154  1.41  8.73  2.31  339.