Convert data into case_data

Convert data into case_data

Usage

as_case_data(
  data,
  id_var = "index_id",
  biomarker_var = "antigen_iso",
  value_var = "value",
  time_in_days = "timeindays"
)

Arguments

data

a data.frame

id_var

a character string naming the column in data denoting participant ID

biomarker_var

a character string naming the column in data denoting which biomarker is being reported in value_var (e.g. "antigen_iso")

value_var

a character string naming the colunmn in data with biomarker measurements

time_in_days

a character string naming the column in data with elapsed time since seroconversion

Value

a case_data object

Examples

sim_case_data() |> 
   as_case_data(
     id_var = "index_id",
     biomarker_var = "antigen_iso",
     time_in_days = "timeindays",
     value_var = "value"
   ) 
#> Error: argument "n" is missing, with no default