Prepare priors

Takes a list input to allow for modifiable priors. Priors can be specified as an option in run_mod.

Usage

prep_priors(max_antigens, priors = NA)

Arguments

max_antigens

integer: how many antigen-isotypes will be modeled

priors

a list with optional elements:

"mu_hyp_param": A matrix of hyperpriors for parameter fixed effects. If specified must be 5 values long.

y0 = baseline antibody concentration (default = 1.0)

y1 = peak antibody concentration (default = 7.0)

t1 = time to peak (default = 1.0)

r = shape parameter (default = -4.0)

alpha = decay rate (default = -1.0)

"prec_hyp_param": Corresponds to diagonal entries representing the precision matrix of hyper-parameter for person-specific fixed effects. Creates an array with 5 matrix, each with 2 x 5 values. If specified must be 5 values long:

defaults: y0 = 1.0, y1 = 0.00001, t1 = 1.0, r = 0.001, alpha = 1.0

"omega_param": Corresponds to diagonal entries representing the precision matrix of Wishart hyper-priors for person-specific random effects. Creates an array with 5 matrix, each with 2 x 5 values. If specified must be 5 values long:

defaults: y0 = 1.0, y1 = 50.0, t1 = 1.0, r = 10.0, alpha = 1.0

"wishdf_param": A vector specifying Wishart distribution degrees of freedom. Creates a vector measuring 1 x 2. If specified, must be 1 value long.

default = 20.0

"prec_logy_hyp_param": A matrix of hyper-parameters for the precision. (inverse variance) of the biomarkers, on log-scale. If specified, must be 2 values long.

defaults = 4.0, 1.0

Value

A list with elements: "n_params": how many parameters??

"mu.hyp": A matrix of hyperpriors for y0, y1, t1, r, and alpha.

"prec.hyp": A three-dimensional numeric array containing the "scale matrix" hyper-parameters for precision corresponding to mu_hyp_param.

"omega" : A three-dimensional numeric array containing the "scale matrix" hyper-parameters of the Wishart hyper-priors on the person-specific random effects, for each antigen-isotype. The first dimension corresponds to the antigen isotypes and has length equal to max_antigens, and the latter two dimensions correspond to the model parameters and each have length equal to n_params.

"wishdf": A vector specifying Wishart distribution degrees of freedom.

"prec.logy.hyp": A matrix of hyper-parameters for the precision (inverse variance) of the biomarkers, on log-scale.

Examples

prep_priors(max_antigens = 2)
#> $n_params
#> [1] 5
#> 
#> $mu.hyp
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    7    1   -4   -1
#> [2,]    1    7    1   -4   -1
#> 
#> $prec.hyp
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    1    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1]  [,2] [,3] [,4] [,5]
#> [1,]    0 1e-05    0    0    0
#> [2,]    0 1e-05    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    0    0
#> [2,]    0    0    1    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]  [,4] [,5]
#> [1,]    0    0    0 0.001    0
#> [2,]    0    0    0 0.001    0
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    1
#> [2,]    0    0    0    0    1
#> 
#> 
#> $omega
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    1    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0   50    0    0    0
#> [2,]    0   50    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    0    0
#> [2,]    0    0    1    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0   10    0
#> [2,]    0    0    0   10    0
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    1
#> [2,]    0    0    0    0    1
#> 
#> 
#> $wishdf
#> [1] 20 20
#> 
#> $prec.logy.hyp
#>      [,1] [,2]
#> [1,]    4    1
#> [2,]    4    1
#> 
#> attr(,"class")
#> [1] "curve_params_priors" "list"               
#> attr(,"used_priors")
#> attr(,"used_priors")$mu_hyp_param
#> [1]  1  7  1 -4 -1
#> 
#> attr(,"used_priors")$prec_hyp_param
#> [1] 1e+00 1e-05 1e+00 1e-03 1e+00
#> 
#> attr(,"used_priors")$omega_param
#> [1]  1 50  1 10  1
#> 
#> attr(,"used_priors")$wishdf_param
#> [1] 20
#> 
#> attr(,"used_priors")$prec_logy_hyp_param
#> [1] 4 1
#>